+++ components/fetch-build-scan-data-cmdline-tool/src/main/java/com/gradle/enterprise/network/NetworkSettingsConfigurator.java
@@ -29,7 +29,7 @@ public class NetworkSettingsConfigurator {
     }
 
     private static boolean isNetworkProperty(String key) {
-        return isSslProperty(key) || isProxyProperty(key);
+        return isSslProperty(key) || isProxyProperty(key) || isConnectionProperty(key);
     }
     private static boolean isSslProperty(String key) {
         return key.startsWith("javax.net.ssl")
@@ -41,6 +41,10 @@ public class NetworkSettingsConfigurator {
             || key.startsWith("https.proxy")
             || key.startsWith("socksProxy")
             || key.endsWith(".nonProxyHosts");
+    }
+
+    private static boolean isConnectionProperty(String key) {
+        return key.startsWith("connect") || key.startsWith("read");
     }
 
     private static Properties loadProperties(Path propertiesFile) throws IOException {