+++ release/changes.md
@@ -2,3 +2,5 @@
 - [NEW] Include serialization factor in all experiment summaries
 - [NEW] Support `-x` command line option for Gradle experiment 1
 - [FIX] Incorrect parsing of additional argument values containing spaces
+- [NEW] Add configuration settings to control the connect and read timeouts when fetching Build Scan data
+